Update1 StatsUnitTest.cs
1. Change from 

            const floatmaxThreshold = 10.2F; ->  const double maxThreshold = 10.2;

+++ Statistics.Test/StatsUnitTest.cs
@@ -35,7 +35,7 @@ namespace Statistics.Test
             var ledAlert = new LEDAlert();
             IAlerter[] alerters = {emailAlert, ledAlert};
 
-            const float maxThreshold = 10.2F;
+            const double maxThreshold = 10.2;
             var statsAlerter = new StatsAlerter(maxThreshold, alerters);
             statsAlerter.checkAndAlert(new List<double>{0.2, 11.9, 4.3, 8.5});
